requirements.txt
Django==1.10.3

django-bootstrap3==7.1.0

django-debug-toolbar==1.6

django-precise-bbcode==1.2.6

django-taggit==0.21.3

easy-thumbnails==2.3

Pillow==3.4.2

